requirements.txt
streamlit==1.46.0
validators==0.28.1
python-dotenv==1.1.1
torch==2.7.1
openai-whisper==20250625
yt-dlp==2025.6.30
langchain==0.3.26
langchain-groq==0.3.5
langchain-community==0.3.27
langchain-core==0.3.68
numpy==2.2.6

